+++ src/renderer/src/components/settings/general-search.ts
@@ -60,37 +60,6 @@ export const GENERAL_CACHE_TIMER_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   }
 ]
 
-export const GENERAL_BROWSER_SEARCH_ENTRIES: SettingsSearchEntry[] = [
-  {
-    title: 'Default Home Page',
-    description: 'URL opened when creating a new browser tab. Leave empty to open a blank tab.',
-    keywords: ['browser', 'home', 'homepage', 'default', 'url', 'new tab', 'blank', 'landing']
-  },
-  {
-    title: 'Terminal Link Routing',
-    description:
-      'Cmd/Ctrl+click opens terminal http(s) links in Orca. Shift+Cmd/Ctrl+click uses the system browser.',
-    keywords: ['browser', 'preview', 'links', 'localhost', 'webview', 'shift', 'cmd', 'ctrl']
-  },
-  {
-    title: 'Session & Cookies',
-    description:
-      'Import cookies from Chrome, Edge, or other browsers to use existing logins inside Orca.',
-    keywords: [
-      'browser',
-      'cookies',
-      'session',
-      'import',
-      'auth',
-      'login',
-      'chrome',
-      'edge',
-      'arc',
-      'profile'
-    ]
-  }
-]
-
 export const GENERAL_CODEX_ACCOUNTS_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   {
     title: 'Codex Accounts',
@@ -114,7 +83,6 @@ export const GENERAL_AGENT_SEARCH_ENTRIES: SettingsSearchEntry[] = [
 
 export const GENERAL_PANE_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   ...GENERAL_WORKSPACE_SEARCH_ENTRIES,
-  ...GENERAL_BROWSER_SEARCH_ENTRIES,
   ...GENERAL_EDITOR_SEARCH_ENTRIES,
   ...GENERAL_CLI_SEARCH_ENTRIES,
   ...GENERAL_CACHE_TIMER_SEARCH_ENTRIES,
